+++ src/features/create-paper/index.ts
@@ -1,0 +1,1 @@
+export { default as CreatePaperForm } from "./ui/create-paper-form"
