+++ src/features/create-paper/ui/background-card.tsx
