+++ src/features/create-paper/ui/background-item.tsx
